add internal functions

+++ R/utils_matrix.R
@@ -27,7 +27,7 @@ from_matrix <- function(x) {
   x_tbl[[.grid_row_name]] <- as.character(1:9)
   x_tbl <- x_tbl %>% 
     pivot_longer(matches("[[:digit:]]"), names_to = .grid_column_name, values_to = .grid_value_name) %>% 
-    mutate_at(c(.grid_row_name, .grid_column_name), as.numeric)
+    mutate_at(c(.grid_row_name, .grid_column_name), as.integer)
   x_tbl[[.grid_given_name]] <- ifelse(is.na(x_tbl[[.grid_value_name]]), NA, TRUE)
   new("SudokuGrid", x_tbl)
 }
